requirements.txt
python-telegram-bot==21.1.1
gspread==6.0.2
google-auth==2.24.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
requests==2.32.3
urllib3==2.3.0
spacy==3.7.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
